project/agents/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
crewai
google-api-python-client>=2.0.0
instagram-private-api>=1.6.0
requests>=2.31.0
python-jose>=3.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.26.2
scikit-learn==1.3.2
groq-sdk>=0.4.0 